moves filesystem properties from treestyle to its own type and makes it a property of treestyle.
